java 공부 - 2024-07-11 업로드

+++ java_hello4/src/OperatePromotion.java
@@ -2,9 +2,10 @@
 public class OperatePromotion {
 
 	public static void main(String[] args) {
-		short num1 = 11;
-		short num2 = 22;
-		short result = num1 + num2;
+		int num1 = 11;
+		int num2 = 22;
+		
+		int result = num1 + num2;
 		System.out.println(result);
 
 	}
